projects/pipeline/requirements.txt
google-cloud-bigquery>=3.17.0
